refactor(platform): move SSRF safe-client module out of catalog
The repo-wide SSRF chokepoint lived at modules/catalog/sources/security.py,
which put two codified rules in direct conflict.

AGENTS.md Rule 2 makes make_safe_client() the mandatory constructor for any
httpx client that follows redirects. Nineteen files import it, including
auth/oauth, platform/config_ops, and five modules under processing/ingest. The
layering burndown in tests/test_layering.py listed
app.modules.catalog.sources.security as shrink-only debt for four of those
processing files, telling the next person who touched them to route the call
through ProcessingPort. That indirection would stop the Rule 2 pre-commit grep
from matching at the import site, so satisfying one rule meant breaking the
other.

The module belongs under platform/. It imports nothing but the standard library
and httpx, and every layer that uses it already sits above platform. Moving it
there leaves both rules intact: the four burndown entries are dead edges now and
are deleted, and the hook still matches.

Module content is byte-identical, only the path changed. There is no re-export
shim at the old path, since one unambiguous home is the point of the move.

Path-pinned references updated: the pre-commit hook (name, guidance message, and
the exclude regex that keeps the factory from tripping its own rule), AGENTS.md
Rule 2, the burndown list, the tiles/router.py forward reference, and the
cog_info.py dual-gate docstring.
