+++ AGENTS.md
@@ -109,7 +109,9 @@ Reference implementations:
 
 **Rule 2 — SSRF redirect-revalidation**
 
-Any new `httpx.AsyncClient` configured with `follow_redirects=True` MUST be constructed via `make_safe_client()` from `backend/app/modules/catalog/sources/security.py` — never directly with `httpx.AsyncClient(follow_redirects=True, ...)`. The factory installs the per-hop `_revalidate_redirect` event hook that re-runs `validate_url_for_ssrf` against every 3xx `Location` header.
+Any new `httpx.AsyncClient` configured with `follow_redirects=True` MUST be constructed via `make_safe_client()` from `backend/app/platform/security.py` — never directly with `httpx.AsyncClient(follow_redirects=True, ...)`. The factory installs the per-hop `_revalidate_redirect` event hook that re-runs `validate_url_for_ssrf` against every 3xx `Location` header.
+
+The module belongs in `platform/` and must stay there. It is cross-cutting infrastructure that auth, config_ops, catalog, and processing all depend on, and it contains no catalog logic. While it lived at `modules/catalog/sources/security.py` this rule contradicted the layering burndown installed by #435, which listed each `processing/` importer as debt to be routed through `ProcessingPort` — the indirection that would stop the Rule 2 grep hook from matching. Do not move it back under a product domain.
 
 GDAL and ogr2ogr CANNOT be made redirect-safe from the inside: `GDAL_HTTP_FOLLOWLOCATION` is not a GDAL option (setting it does nothing — #937), and GDAL exposes no option that disables redirect-following. For any GDAL/ogr2ogr/rasterio path the defenses are structural, in this order: (1) prefer never handing a caller-controlled URL to GDAL at all — fetch only managed storage (`/vsis3/`, `/vsiaz/`, local paths with validated keys) and never probe remote sources in-process; (2) where a user-supplied service URL must be fetched (service ingest/preview), `validate_url_for_ssrf` gates it at submission time and residual redirect/DNS-rebinding exposure is bounded operationally (worker egress firewall); (3) subprocess envs come from `gdal_safe_env()` / `gdal_safe_open_env()` in `backend/app/processing/raster/vrt.py`, which apply the real clamps (`CPL_VSIL_CURL_ALLOWED_EXTENSIONS`, `VRT_VIRTUAL_OVERVIEWS`). Never re-add `GDAL_HTTP_FOLLOWLOCATION` anywhere; it reads as a defense and is a no-op.
