+++ backend/tests/test_reupload_service.py
@@ -227,7 +227,7 @@ class TestServiceReuploadWorker:
             # SSRF re-validation of ``source_url`` at the top of the
             # ``reupload_service`` worker body (tasks_reupload.py:373-378)
             # via a LAZY from-import inside the function body:
-            #     from app.modules.catalog.sources.security import validate_url_for_ssrf
+            #     from app.platform.security import validate_url_for_ssrf
             # The lazy import re-binds the symbol on every call, so the
             # patch target MUST be the function's defining module (NOT
             # the worker's namespace). The fixture URL
@@ -237,7 +237,7 @@ class TestServiceReuploadWorker:
             # the gate via AsyncMock. Same fix shape as Plan 1075-03
             # closed in test_ingest.py:1369-1372.
             patch(
-                "app.modules.catalog.sources.security.validate_url_for_ssrf",
+                "app.platform.security.validate_url_for_ssrf",
                 new=AsyncMock(),
             ),
             patch(
@@ -383,7 +383,7 @@ class TestServiceReuploadWorker:
             # contract (401-retry guidance message from IngestionError).
             # Plan 1075-03 / test_ingest.py:1369 is the canonical pattern.
             patch(
-                "app.modules.catalog.sources.security.validate_url_for_ssrf",
+                "app.platform.security.validate_url_for_ssrf",
                 new=AsyncMock(),
             ),
             patch(
